backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
facenet-pytorch
multipart
